# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Accountability PoC",
    "description": "PoC backend for timed sessions, RAG MCQ prompts, and a simple leaderboard. See `/docs` for OpenAPI UI.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/leaderboard/entries": {
      "post": {
        "tags": [
          "leaderboard"
        ],
        "summary": "Create raw leaderboard entry",
        "description": "Create a raw leaderboard entry record (for analytics or ingest).",
        "operationId": "create_entry_leaderboard_entries_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "score",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Score"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "leaderboard"
        ],
        "summary": "List raw entries",
        "description": "List raw leaderboard entries sorted by score (descending).",
        "operationId": "list_entries_leaderboard_entries_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardEntry"
                  },
                  "title": "Response List Entries Leaderboard Entries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leaderboard/entries/me": {
      "get": {
        "tags": [
          "leaderboard"
        ],
        "summary": "My score history",
        "description": "Return all LeaderboardEntry audit rows for the current user, newest first.",
        "operationId": "my_entries_leaderboard_entries_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardEntry"
                  },
                  "title": "Response My Entries Leaderboard Entries Me Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leaderboard/aggregate/top": {
      "get": {
        "tags": [
          "leaderboard"
        ],
        "summary": "Top raw entries",
        "description": "Return top raw entries by score.",
        "operationId": "top_by_user_leaderboard_aggregate_top_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardEntry"
                  },
                  "title": "Response Top By User Leaderboard Aggregate Top Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leaderboard/submit": {
      "post": {
        "tags": [
          "leaderboard"
        ],
        "summary": "Submit session result",
        "description": "Submit a session result. This performs an upsert to the per-user aggregate (best/total/sessions).",
        "operationId": "submit_result_leaderboard_submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "score",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Score"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardAggregate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leaderboard/aggregates": {
      "get": {
        "tags": [
          "leaderboard"
        ],
        "summary": "List aggregates",
        "description": "List per-user aggregates; order by `best` (default) or `total`.",
        "operationId": "list_aggregates_leaderboard_aggregates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "best",
              "title": "Order By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardAggregate"
                  },
                  "title": "Response List Aggregates Leaderboard Aggregates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leaderboard/aggregate/{user_id}": {
      "get": {
        "tags": [
          "leaderboard"
        ],
        "summary": "Get user aggregate",
        "description": "Fetch the per-user aggregate (uses a short in-process TTL cache). Returns 404 if missing.",
        "operationId": "get_aggregate_leaderboard_aggregate__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardAggregate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leaderboard/reset": {
      "post": {
        "tags": [
          "leaderboard"
        ],
        "summary": "Admin: reset leaderboard",
        "description": "Admin-only: clear leaderboard tables and cache. Requires `ADMIN_TOKEN` env var to be set and passed in X-Admin-Token header.",
        "operationId": "reset_leaderboard_leaderboard_reset_post",
        "parameters": [
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeaderboardAggregate": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "best_score": {
            "type": "number",
            "title": "Best Score",
            "default": 0.0
          },
          "total_score": {
            "type": "number",
            "title": "Total Score",
            "default": 0.0
          },
          "sessions": {
            "type": "integer",
            "title": "Sessions",
            "default": 0
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "LeaderboardAggregate"
      },
      "LeaderboardEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "session_count": {
            "type": "integer",
            "title": "Session Count",
            "default": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "score"
        ],
        "title": "LeaderboardEntry",
        "description": "Audit-trail row: one per finished session."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}